Move JsonUtils to util package
